Added support for parsing structs declared inside other structs (nested definitions)
Note that the reason why it did not work before was due to a bug in castxml. The
XML output format was changed from --castxml-output=1 to --castxml-gccxml. This
might have unknown negative (or positive) side effects. However, all xml files in
the test directory has been converted to the new format and tests pass successfully.